+++ account_fiscal_position_payable_receivable/__init__.py
@@ -1,3 +1,1 @@
-# -*- coding: utf-8 -*-
-
-from . import account_fiscal_position
+from . import models
