+++ account_fiscal_position_payable_receivable/__manifest__.py
@@ -1,10 +1,9 @@
-# -*- coding: utf-8 -*-
 # © 2016-2017 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 {
     'name': 'Account Fiscal Position Payable Receivable',
-    'version': '10.0.1.0.0',
+    'version': '12.0.1.0.0',
     'category': 'Accounting & Finance',
     'license': 'AGPL-3',
     'summary': 'Configure payable/receivable accounts on fiscal positions',
@@ -20,7 +19,7 @@ This module has been written by Alexis de Lattre from Akretion <alexis.delattre@
     'website': 'http://www.akretion.com',
     'depends': ['account'],
     'data': [
-        'account_fiscal_position_view.xml',
+        'views/account_fiscal_position_view.xml',
     ],
     'installable': True,
 }
