+++ account_fiscal_position_payable_receivable/models/__init__.py
@@ -1,0 +1,2 @@
+from . import account_fiscal_position
+from . import res_partner
